requirements.txt
langchain==0.1.0
python-dotenv==1.0.0
chromadb==0.4.16
unstructured==0.16.25
PyPDF2==3.0.1
google-generativeai==0.3.1
